Added an url validation for the 'set' command

+++ src/commands/set.js
@@ -1,5 +1,7 @@
+const validator = require('validator');
 const { Sound } = require("../models");
 const { Command } = require("../core/command");
+const { CommandError } = require('../core/errors');
 
 const command = new Command({ name: "set" });
 
@@ -11,14 +13,16 @@ command.execute = async function (message, args) {
     throw new Error("!soundId || !soundUrl");
   };
 
+  if (!validator.isURL(soundUrl)) {
+    throw new CommandError("Stop trolling that's not a url");
+  }
+
   await Sound.destroy({
-    where: {
-      guildId,
-      soundId,
-    },
+    where: { guildId, soundId },
   });
 
   await Sound.create({ guildId, soundId, soundUrl });
 }
 
 module.exports = command;
+
